requirements.txt
mysql-connector-python==8.3.0
sqlparse==0.4.4
cryptography==42.0.5